# docs/swagger.yaml
basePath: /api
definitions:
  auth.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  auth.userResponse:
    properties:
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  connection.connectionResponse:
    properties:
      createdAt:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      lastCheckedAt:
        type: string
      maskedApiKey:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
  connection.createConnectionRequest:
    properties:
      apiKey:
        type: string
      name:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  connection.updateConnectionRequest:
    properties:
      apiKey:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for media-reaper, a tool that integrates Sonarr/Radarr with Emby
    to identify watched media eligible for deletion.
  title: Media Reaper API
  version: 0.1.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with username and password to create a session
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Destroy the current session and clear the session cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Get the currently authenticated user from the session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.userResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionCookie: []
      summary: Current user
      tags:
      - auth
  /connections:
    get:
      description: List all connections with masked API keys
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/connection.connectionResponse'
            type: array
      security:
      - SessionCookie: []
      summary: List connections
      tags:
      - connections
    post:
      consumes:
      - application/json
      description: Create a new Sonarr, Radarr, or Emby connection with encrypted
        API key
      parameters:
      - description: Connection details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/connection.createConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/connection.connectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionCookie: []
      summary: Create connection
      tags:
      - connections
  /connections/{id}:
    delete:
      description: Delete a connection by ID
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionCookie: []
      summary: Delete connection
      tags:
      - connections
    get:
      description: Get a connection by ID with masked API key
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/connection.connectionResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionCookie: []
      summary: Get connection
      tags:
      - connections
    put:
      consumes:
      - application/json
      description: Update a connection. API key is optional (omit to keep current).
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: string
      - description: Connection details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/connection.updateConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/connection.connectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionCookie: []
      summary: Update connection
      tags:
      - connections
  /health:
    get:
      description: Returns service health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - system
securityDefinitions:
  SessionCookie:
    in: cookie
    name: media-reaper-session
    type: apiKey
swagger: "2.0"